termino dos graficos e divisão dos arquivos js
