+++ html/login.html
@@ -1,5 +1,6 @@
 <!DOCTYPE html>
 <html lang="pt-BR">
+
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -9,16 +10,17 @@
   <link rel="shortcut icon" href="img/ecomanager.png" type="image/x-icon">
   <title>Login - EcoManager</title>
 </head>
+
 <body>
   <script type="module">
-  // Import the functions you need from the SDKs you need
-  import { initializeApp } from "https://www.gstatic.com/firebasejs/12.1.0/firebase-app.js";
-  import { getAnalytics } from "https://www.gstatic.com/firebasejs/12.1.0/firebase-analytics.js";
-  // TODO: Add SDKs for Firebase products that you want to use
-  // https://firebase.google.com/docs/web/setup#available-libraries
+    // Import the functions you need from the SDKs you need
+    import { initializeApp } from "https://www.gstatic.com/firebasejs/12.1.0/firebase-app.js";
+    import { getAnalytics } from "https://www.gstatic.com/firebasejs/12.1.0/firebase-analytics.js";
+    // TODO: Add SDKs for Firebase products that you want to use
+    // https://firebase.google.com/docs/web/setup#available-libraries
 
-  // Your web app's Firebase configuration
-  // For Firebase JS SDK v7.20.0 and later, measurementId is optional
+    // Your web app's Firebase configuration
+    // For Firebase JS SDK v7.20.0 and later, measurementId is optional
     const firebaseConfig = {
       apiKey: "AIzaSyD1OyAxSZWLQq_uaMQJzFxxactW-sMLZ4M",
       authDomain: "ecomanager-tcc.firebaseapp.com",
@@ -29,19 +31,19 @@
       measurementId: "G-6TTFT5PVCX"
     };
 
-  // Initialize Firebase
-  const app = initializeApp(firebaseConfig);
-  const analytics = getAnalytics(app);
-</script>
+    // Initialize Firebase
+    const app = initializeApp(firebaseConfig);
+    const analytics = getAnalytics(app);
+  </script>
 
-  
+
 
   <div class="login-container">
     <div class="login-box">
       <img src="img/logo-nome.png" alt="EcoManager Logo" class="login-logo">
       <h2>Acesse a Plataforma</h2>
       <p>Use sua conta preferida para aceder aos dashboards e calculadoras de forma segura.</p>
-      
+
       <button id="google-login-btn" class="btn-provider btn-google">
         <img src="https://upload.wikimedia.org/wikipedia/commons/c/c1/Google_%22G%22_logo.svg" alt="Google Logo">
         Entrar com Google
@@ -63,8 +65,9 @@
   <!-- Scripts do Firebase -->
   <script src="https://www.gstatic.com/firebasejs/8.10.0/firebase-app.js"></script>
   <script src="https://www.gstatic.com/firebasejs/8.10.0/firebase-auth.js"></script>
-  
+
   <!-- Nosso script de login -->
   <script src="js/login.js"></script>
 </body>
+
 </html>
