+++ js/scriptCalculadoras/tabs.js
@@ -9,6 +9,20 @@ export function initCalculatorTabs() {
       tabContents.forEach(content => content.classList.remove("active"));
       button.classList.add("active");
       document.getElementById(targetTab)?.classList.add("active");
+
+      tabContents.forEach(content => {
+        if (content.id !== targetTab) {
+          limparCampos(content);
+        }
+      });
     });
+  });
+}
+
+function limparCampos(container) {
+  const inputs = container.querySelectorAll("input, select, textarea");
+  inputs.forEach(input => {
+    if (input.id === 'tarifa' || input.id === 'agua-tarifa') return;
+    input.value = "";
   });
 }
